+++ package.json
@@ -12,7 +12,7 @@
    "license": "ISC",
    "dependencies": {
       "cors": "^2.8.5",
-      "dayjs": "^1.10.3",
+      "dayjs": "^1.10.4",
       "dotenv": "^8.2.0",
       "express": "4.17.0",
       "knex": "^0.21.1",
